+++ sdk/cosmos/azure_data_cosmos/tests/emulator_tests/cosmos_query.rs
@@ -6,12 +6,13 @@ use super::framework;
 
 use std::error::Error;
 
+use azure_data_cosmos::feed::ContinuationToken;
 use azure_data_cosmos::{
     clients::DatabaseClient,
+    feed::FeedScope,
     models::CosmosStatus,
     options::{MaxItemCountHint, QueryOptions},
-    query::FeedScope,
-    ContinuationToken, Query,
+    Query,
 };
 use framework::{test_data, MockItem, TestClient};
 use futures::StreamExt;